+++ app/Http/Controllers/RolesController.php
@@ -9,9 +9,13 @@ use Validator;
 class RolesController extends Controller
 {
     public function index(){
+        if(dsld_have_user_permission('roles') == 0){
+            return redirect()->route('backend.admin')->with('error', 'You have no permission');
+        }
         $page['title'] = 'Show all role';
         return view('backend.modules.roles.show', compact('page'));
     }
+
     public function get_ajax_roles(Request $request){
         if($request->page != 1){$start = $request->page * 25;}else{$start = 0;}
         $search = $request->search;
@@ -38,8 +42,11 @@ class RolesController extends Controller
         $data = $data->skip($start)->paginate(25);
         return view('backend.modules.roles.ajax_roles', compact('data'));
     }
-    public function store(Request $request){
 
+    public function store(Request $request){
+        if(dsld_have_user_permission('roles_add') == 0){
+            return response()->json(['status' => 'error', 'message'=> "You have no permission."]);
+        }
         $validator = Validator::make($request->all(), [
             'name' => 'required|string|max:50',
         ]);
@@ -63,14 +70,22 @@ class RolesController extends Controller
             return response()->json(['status' => 'warning', 'message'=> 'Details already exist! please try agin.']);
         }
     }
+
     public function edit($id){
+        if(dsld_have_user_permission('roles_edit') == 0){
+            return redirect()->route('backend.admin')->with('error', 'You have no permission');
+        }
+        
         $data = Role::where('id', $id)->first();
         $page['title'] = 'Edit Data';
         return view('backend.modules.roles.edit', compact('data', 'page'));
     }
     
     public function update(Request $request){
-
+        if(dsld_have_user_permission('roles_edit') == 0){
+            return response()->json(['status' => 'error', 'message'=> "You have no permission."]);
+        }
+        
         $validator = Validator::make($request->all(), [
             'name' => 'required|string|max:50',
         ]);
@@ -96,7 +111,11 @@ class RolesController extends Controller
         }
 
     }
+    
     public function destory(Request $request){
+        if(dsld_have_user_permission('roles_delete') == 0){
+            return response()->json(['status' => 'error', 'message'=> "You have no permission."]);
+        }
         $user = Role::findOrFail($request->id);
         if($user != ''){
             if($user->delete()){
